+++ CHANGELOG.md
@@ -33,9 +33,11 @@ with `this` context of the response.
 passed `req` and `res` in addition to `err`.
 - `status()` assertion method asserts that object is an instance of `Response`,
 rather than just having a `statusCode` property.
+- `Response#setHeader` returns response object to match behavior of
+`http.ServerResponse`.
 - Changed `Response#redirect` signature from `function(url, status)` to
 `function(status, url)`, matching Express 4.x API.
-- `Response#status()` added to `res` in `init()` middleware rather than being
+- `Response#status` added to `res` in `init()` middleware rather than being
 part of base `Response` prototype.
 
 ### Removed